Float variable as reduction size for imresize.c

+++ src/CControl/Headers/Functions.h
@@ -161,7 +161,7 @@ FAST_XY* fast(const uint8_t X[], int row, int column, int threshold, int* num_co
 size_t hough(float X[], float* K[], float* M[], float p, float epsilon, size_t min_pts, size_t row, size_t column);
 void imshow(PGM* image);
 PGM* imread(const char file_path[]);
-void imresize(float X[], float Y[], size_t d, size_t row, size_t column);
+void imresize(float X[], float Y[], float d, size_t row, size_t column);
 void imgaussfilt(float X[], float sigma, size_t row, size_t column);
 void imfree(PGM* image);
 void pooling(float A[], float P[], size_t row_a, size_t column_a, size_t p, POOLING_METHOD pooling_method);
